rename clmm/utils/tools.py -> clmm/utils/validation.py
